requirements.txt
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
PyYAML==6.0.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
